qr기능-2

+++ api/route/quest.py
@@ -199,42 +199,50 @@ def app_check(teacher_id, id):
             "msg": "요청 오류",
         })
 
+
 @bp.route("/game/<string:user_id>", methods=["GET", "POST"])
 @login_required
 def game_quest(user_id): # 자신의 퀘스트 목록 가져오기(GET), 퀘스트 완료요청 보내기(POST)
-    if request.method == "GET":
-        now = datetime.datetime.now()
-        user_quest = UserQuest.query.filter(and_(UserQuest.user_id == user_id, UserQuest.start_date <= now <= UserQuest.end_date))
-        user_quest = serializable_userQuest(user_quest)
-        db.session.remove()
-        return jsonify({
-            "code": 1,
-            "msg": "유저 퀘스트 목록 반환!",
-            "data": user_quest
-        })
-    elif request.method == "POST": # 퀘스트 완료 요청
-        user_quest = UserQuest.query.filter(and_(UserQuest.user_id == user_id, UserQuest.id == request.json['id'])).first()
-        now = datetime.datetime.now()
-
-        if user_quest.create_date > now:
-            return jsonify({
-                "code": -1,
-                "msg": "퀘스트가 아직 활성화되지 않았습니다.",
-            })
-        elif user_quest.end_date < now:
-            db.session.delete(user_quest)
+    user = User.query.filter_by(user_id = user_id).first()
+    if user.isStudent == True:
+        if request.method == "GET":
+            now = datetime.datetime.now()
+            user_quest = UserQuest.query.filter(and_(UserQuest.user_id == user_id, UserQuest.start_date <= now <= UserQuest.end_date))
+            user_quest = serializable_userQuest(user_quest)
             db.session.remove()
             return jsonify({
-            "code": -1,
-            "msg": "퀘스트 기간이 만료되었습니다.",
+                "code": 1,
+                "msg": "유저 퀘스트 목록 반환!",
+                "data": user_quest
             })
-        else:
-            user_quest.done = True
-        db.session.commit()
-        db.session.remove()
+        elif request.method == "POST": # 퀘스트 완료 요청
+            user_quest = UserQuest.query.filter(and_(UserQuest.user_id == user_id, UserQuest.id == request.json['id'])).first()
+            now = datetime.datetime.now()
+
+            if user_quest.create_date > now:
+                return jsonify({
+                    "code": -1,
+                    "msg": "퀘스트가 아직 활성화되지 않았습니다.",
+                })
+            elif user_quest.end_date < now:
+                db.session.delete(user_quest)
+                db.session.remove()
+                return jsonify({
+                "code": -1,
+                "msg": "퀘스트 기간이 만료되었습니다.",
+                })
+            else:
+                user_quest.done = True
+            db.session.commit()
+            db.session.remove()
+            return jsonify({
+                "code": 1,
+                "msg": "퀘스트 완료 요청을 보냈습니다.",
+            })
+    else:
         return jsonify({
-            "code": 1,
-            "msg": "퀘스트 완료 요청을 보냈습니다.",
+            "code" : -1,
+            "msg" : "학생만 이용가능합니다."
         })
